# internal/entity/user.go
package entity

import (
	"errors"
	"time"

	"github.com/golang-jwt/jwt/v5"
)

var (
	ErrUserNotFound               = errors.New("user not found")
	ErrAuthorizationFailed        = errors.New("authorization failed")
	ErrMissingAuthorizationHeader = errors.New("missing Authorization header")
	ErrTokenExpired               = errors.New("token expired")
	ErrInvalidToken               = errors.New("invalid token")
	ErrFailedToReadSecret         = errors.New("failed to read secret")
	ErrFailedToGetSecretVersion   = errors.New("failed to get secret version")
	ErrFailedToReadSignature      = errors.New("failed to read signature")
)

type UserID string

func (id UserID) String() string {
	return string(id)
}

type User struct {
	ID           UserID
	Name         string
	Email        string
	PasswordHash string
	CreatedAt    time.Time
	UpdatedAt    time.Time
}

type UserCreate struct {
	Name            string `validate:"required"`
	Email           string `validate:"required,email"`
	Password        string `validate:"required,min=12,max=128"`
	PasswordConfirm string `validate:"required,eqfield=Password"`
}

type UserLogin struct {
	Email    string `validate:"required,email"`
	Password string `validate:"required"`
}

type UserToken struct {
	Token            string
	RefreshToken     string
	ExpiresAt        time.Time
	RefreshExpiresAt time.Time
}

type RefreshTokenRequest struct {
	RefreshToken string `validate:"required,jwt"`
}

type UserClaims struct {
	Version string `json:"ver"`
	jwt.RegisteredClaims
}

type VerifyToken struct {
	Token string `validate:"required,jwt"`
}

type ChangePasswordRequest struct {
	Email              string `validate:"required,email"`
	CurrentPassword    string `validate:"required"`
	NewPassword        string `validate:"required,min=12,max=128"`
	NewPasswordConfirm string `validate:"required,eqfield=NewPassword"`
}

type PasswordUpdate struct {
	Email           string
	NewPasswordHash string
}